Imported 35.1.0-5.0.2 source

+++ kernel/nvgpu/drivers/gpu/nvgpu/common/cic/mon/mon_lut.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2021-2022, NVIDIA CORPORATION.  All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -96,5 +96,5 @@ int nvgpu_cic_mon_get_num_hw_modules(struct gk20a *g)
 		return -EINVAL;
 	}
 
-	return g->cic_mon->num_hw_modules;
+	return (int)g->cic_mon->num_hw_modules;
 }
