+++ kernel/nvgpu/drivers/gpu/nvgpu/hal/therm/therm_tu104.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2019-2022, NVIDIA CORPORATION.  All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -33,7 +33,7 @@
 void tu104_get_internal_sensor_limits(s32 *max_24_8, s32 *min_24_8)
 {
 	*max_24_8 = (0x87 << 8);
-	*min_24_8 = (((u32)-216) << 8);
+	*min_24_8 = (s32)(((u32)-216) << 8);
 }
 
 void tu104_get_internal_sensor_curr_temp(struct gk20a *g, u32 *temp_f24_8)
